requirements.txt
tensorflow_gpu==1.5.0
scikit_image==0.15.0
scikit_learn==0.21.2
h5py==2.9.0
tqdm==4.32.1
pyemd==0.5.1
Pillow==6.1.0
